Ensure columns used with numeric filters in variable map are ints

+++ mimic_direct_extract.py
@@ -174,11 +174,15 @@ def save_pop(
 def get_variable_mapping(mimic_mapping_filename):
     # Read in the second level mapping of the itemids
     var_map = DataFrame.from_csv(mimic_mapping_filename, index_col=None).fillna('').astype(str)
+    
+    var_map.ITEMID = var_map.ITEMID.astype(int)
+    var_map.COUNT = var_map.COUNT.astype(int)
+    var_map.M_E_APPENDIXB = var_map.M_E_APPENDIXB.astype(int)
+    
     var_map = var_map.ix[(var_map['LEVEL2'] != '') & (var_map.COUNT>0)]
     var_map = var_map.ix[(var_map.STATUS == 'ready')]
     var_map = var_map.ix[(var_map.M_E_APPENDIXB == 1)]
-    var_map.ITEMID = var_map.ITEMID.astype(int)
-
+    
     return var_map
 
 def get_variable_ranges(range_filename):
